backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.55.0
anthropic==0.39.0
supabase==2.10.0
yt-dlp==2024.12.6
pydantic==2.10.0
python-dotenv==1.0.1
moviepy==2.1.1
requests>=2.31.0
# 無料オプション
faster-whisper>=1.0.0
google-genai>=1.0.0
pydub>=0.25.0
